refactor: removing unnecessary authenticated-menu folder
Moving AuthenticatedMenu up a level.
