--- shell/header/authenticated-menu/index.ts
@@ -1,3 +1,0 @@
-import AuthenticatedMenu from './AuthenticatedMenu';
-
-export default AuthenticatedMenu;
